+++ NEWS.md
@@ -1,3 +1,5 @@
+# dRiftDM (development version)
+
 # dRiftDM 0.3.1
 
 ## Minor